+++ components/aos_apps/aos_apps.c
@@ -15,6 +15,7 @@ void aos_apps_register_builtin(void)
         aos_app_music_get,
         aos_app_remote_get,
         aos_app_pcremote_get,
+        aos_app_pato_get,
         aos_app_photos_get,
         aos_app_flashlight_get,
         aos_app_level_get,